+++ 1.vultest/vultest/vultest/Debug/vultest.log
@@ -1,34 +1,28 @@
-﻿生成启动时间为 2017/2/9 8:52:30。
-项目“C:\Users\long\windows_7_x64_pin\1.vultest\vultest\vultest\vultest.vcxproj”在节点 2 上(rebuild 个目标)。
-_PrepareForClean:
-  正在删除文件“Debug\vultest.lastbuildstate”。
+﻿生成启动时间为 2017/2/10 8:38:17。
+项目“C:\Users\long\windows_7_x64_pin\1.vultest\vultest\vultest\vultest.vcxproj”在节点 2 上(build 个目标)。
 InitializeBuildStatus:
   正在创建“Debug\vultest.unsuccessfulbuild”，因为已指定“AlwaysCreate”。
 ClCompile:
-  C:\Program Files\Microsoft Visual Studio 10.0\VC\bin\CL.exe /c /ZI /nologo /W3 /WX- /Od /Oy- /D WIN32 /D _DEBUG /D _CONSOLE /D _UNICODE /D UNICODE /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Yc"StdAfx.h" /Fp"Debug\vultest.pch" /Fo"Debug\\" /Fd"Debug\vc100.pdb" /Gd /TP /analyze- /errorReport:prompt stdafx.cpp
-  stdafx.cpp
+  所有输出均为最新。
   C:\Program Files\Microsoft Visual Studio 10.0\VC\bin\CL.exe /c /ZI /nologo /W3 /WX- /Od /Oy- /D WIN32 /D _DEBUG /D _CONSOLE /D _UNICODE /D UNICODE /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Yu"StdAfx.h" /Fp"Debug\vultest.pch" /Fo"Debug\\" /Fd"Debug\vc100.pdb" /Gd /TP /analyze- /errorReport:prompt vultest.cpp
   vultest.cpp
 ManifestResourceCompile:
-  C:\Program Files\Microsoft SDKs\Windows\v7.0A\bin\rc.exe /nologo /fo"Debug\vultest.exe.embed.manifest.res" Debug\vultest_manifest.rc 
+  所有输出均为最新。
 Link:
   C:\Program Files\Microsoft Visual Studio 10.0\VC\bin\link.exe /ERRORREPORT:PROMPT /OUT:"C:\Users\long\windows_7_x64_pin\1.vultest\vultest\Debug\vultest.exe" /INCREMENTAL /NOLOGO kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /MANIFEST /ManifestFile:"Debug\vultest.exe.intermediate.manifest" /MANIFESTUAC:"level='asInvoker' uiAccess='false'" /DEBUG /PDB:"C:\Users\long\windows_7_x64_pin\1.vultest\vultest\Debug\vultest.pdb" /SUBSYSTEM:CONSOLE /TLBID:1 /DYNAMICBASE /NXCOMPAT /IMPLIB:"C:\Users\long\windows_7_x64_pin\1.vultest\vultest\Debug\vultest.lib" /MACHINE:X86 Debug\vultest.exe.embed.manifest.res
   Debug\stdafx.obj
   Debug\vultest.obj
 Manifest:
-  正在删除文件“Debug\vultest.exe.embed.manifest”。
   C:\Program Files\Microsoft SDKs\Windows\v7.0A\bin\mt.exe /nologo /verbose /out:"Debug\vultest.exe.embed.manifest" /manifest Debug\vultest.exe.intermediate.manifest
-  C:\Program Files\Microsoft SDKs\Windows\v7.0A\bin\rc.exe /nologo /fo"Debug\vultest.exe.embed.manifest.res" Debug\vultest_manifest.rc 
+  所有输出均为最新。
 LinkEmbedManifest:
-  C:\Program Files\Microsoft Visual Studio 10.0\VC\bin\link.exe /ERRORREPORT:PROMPT /OUT:"C:\Users\long\windows_7_x64_pin\1.vultest\vultest\Debug\vultest.exe" /INCREMENTAL /NOLOGO kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /MANIFEST /ManifestFile:"Debug\vultest.exe.intermediate.manifest" /MANIFESTUAC:"level='asInvoker' uiAccess='false'" /DEBUG /PDB:"C:\Users\long\windows_7_x64_pin\1.vultest\vultest\Debug\vultest.pdb" /SUBSYSTEM:CONSOLE /TLBID:1 /DYNAMICBASE /NXCOMPAT /IMPLIB:"C:\Users\long\windows_7_x64_pin\1.vultest\vultest\Debug\vultest.lib" /MACHINE:X86 Debug\vultest.exe.embed.manifest.res
-  Debug\stdafx.obj
-  Debug\vultest.obj
+  所有输出均为最新。
   vultest.vcxproj -> C:\Users\long\windows_7_x64_pin\1.vultest\vultest\Debug\vultest.exe
 FinalizeBuildStatus:
   正在删除文件“Debug\vultest.unsuccessfulbuild”。
   正在对“Debug\vultest.lastbuildstate”执行 Touch 任务。
-已完成生成项目“C:\Users\long\windows_7_x64_pin\1.vultest\vultest\vultest\vultest.vcxproj”(rebuild 个目标)的操作。
+已完成生成项目“C:\Users\long\windows_7_x64_pin\1.vultest\vultest\vultest\vultest.vcxproj”(build 个目标)的操作。
 
 生成成功。
 
-已用时间 00:00:02.16
+已用时间 00:00:03.58
